:truck: :: Move bottom bar
